+++ CHANGELOG.md
@@ -1,3 +1,50 @@
+<a name="v0.51.4"></a>
+## [v0.51.4] - 2022-07-05
+<a name="v0.51.3"></a>
+## [v0.51.3] - 2022-07-05
+<a name="v0.51.2"></a>
+## [v0.51.2] - 2022-07-01
+<a name="v0.51.1"></a>
+## [v0.51.1] - 2022-06-30
+### Bug Fixes
+- **install:** fix serial install detection for windows ([#1326](https://github.com/newrelic/newrelic-client-go/issues/1326))
+
+<a name="v0.51.0"></a>
+## [v0.51.0] - 2022-06-30
+### Features
+- **install:** ensure only 1 newrelic install is running at once
+
+<a name="v0.50.14"></a>
+## [v0.50.14] - 2022-06-29
+<a name="v0.50.13"></a>
+## [v0.50.13] - 2022-06-28
+<a name="v0.50.12"></a>
+## [v0.50.12] - 2022-06-27
+<a name="v0.50.11"></a>
+## [v0.50.11] - 2022-06-27
+<a name="v0.50.10"></a>
+## [v0.50.10] - 2022-06-27
+<a name="v0.50.9"></a>
+## [v0.50.9] - 2022-06-24
+<a name="v0.50.8"></a>
+## [v0.50.8] - 2022-06-23
+<a name="v0.50.7"></a>
+## [v0.50.7] - 2022-06-16
+<a name="v0.50.6"></a>
+## [v0.50.6] - 2022-06-16
+<a name="v0.50.5"></a>
+## [v0.50.5] - 2022-06-14
+<a name="v0.50.4"></a>
+## [v0.50.4] - 2022-06-13
+<a name="v0.50.3"></a>
+## [v0.50.3] - 2022-06-13
+<a name="v0.50.2"></a>
+## [v0.50.2] - 2022-06-08
+<a name="v0.50.1"></a>
+## [v0.50.1] - 2022-06-07
+### Bug Fixes
+- **install:** Remove changelog writing to bypass sidekick error ([#1314](https://github.com/newrelic/newrelic-client-go/issues/1314))
+
 <a name="v0.50.0"></a>
 ## [v0.50.0] - 2022-06-03
 ### Features
@@ -1682,7 +1729,26 @@
 - **profile:** Enable reading of profiles and use Region/APIKey from default profile
 - **profile:** Add listing of profiles to command
 
-[Unreleased]: https://github.com/newrelic/newrelic-client-go/compare/v0.50.0...HEAD
+[Unreleased]: https://github.com/newrelic/newrelic-client-go/compare/v0.51.4...HEAD
+[v0.51.4]: https://github.com/newrelic/newrelic-client-go/compare/v0.51.3...v0.51.4
+[v0.51.3]: https://github.com/newrelic/newrelic-client-go/compare/v0.51.2...v0.51.3
+[v0.51.2]: https://github.com/newrelic/newrelic-client-go/compare/v0.51.1...v0.51.2
+[v0.51.1]: https://github.com/newrelic/newrelic-client-go/compare/v0.51.0...v0.51.1
+[v0.51.0]: https://github.com/newrelic/newrelic-client-go/compare/v0.50.14...v0.51.0
+[v0.50.14]: https://github.com/newrelic/newrelic-client-go/compare/v0.50.13...v0.50.14
+[v0.50.13]: https://github.com/newrelic/newrelic-client-go/compare/v0.50.12...v0.50.13
+[v0.50.12]: https://github.com/newrelic/newrelic-client-go/compare/v0.50.11...v0.50.12
+[v0.50.11]: https://github.com/newrelic/newrelic-client-go/compare/v0.50.10...v0.50.11
+[v0.50.10]: https://github.com/newrelic/newrelic-client-go/compare/v0.50.9...v0.50.10
+[v0.50.9]: https://github.com/newrelic/newrelic-client-go/compare/v0.50.8...v0.50.9
+[v0.50.8]: https://github.com/newrelic/newrelic-client-go/compare/v0.50.7...v0.50.8
+[v0.50.7]: https://github.com/newrelic/newrelic-client-go/compare/v0.50.6...v0.50.7
+[v0.50.6]: https://github.com/newrelic/newrelic-client-go/compare/v0.50.5...v0.50.6
+[v0.50.5]: https://github.com/newrelic/newrelic-client-go/compare/v0.50.4...v0.50.5
+[v0.50.4]: https://github.com/newrelic/newrelic-client-go/compare/v0.50.3...v0.50.4
+[v0.50.3]: https://github.com/newrelic/newrelic-client-go/compare/v0.50.2...v0.50.3
+[v0.50.2]: https://github.com/newrelic/newrelic-client-go/compare/v0.50.1...v0.50.2
+[v0.50.1]: https://github.com/newrelic/newrelic-client-go/compare/v0.50.0...v0.50.1
 [v0.50.0]: https://github.com/newrelic/newrelic-client-go/compare/v0.49.13...v0.50.0
 [v0.49.13]: https://github.com/newrelic/newrelic-client-go/compare/v0.49.12...v0.49.13
 [v0.49.12]: https://github.com/newrelic/newrelic-client-go/compare/v0.49.11...v0.49.12